+++ src/test/kotlin/model/PolynomialTest.kt
@@ -608,7 +608,7 @@ class PolynomialTest {
         with(Polynomial.over(Models.intModP(7))){
             val f = x + x.pow(8) // x + x^8 = x (1 + x^7) = x (1 + x)^7
             val result = f.squareFreeFactorize()
-            assertEquals(f, result.fold(one){acc, pair -> acc * pair.first.pow(pair.second.toLong())})
+            assertEquals(f, result.fold(one){acc, pair -> acc * pair.first.pow(pair.second)})
         }
     }
 //
